+++ examples/scripts/getting_started/maximum_likelihood.py
@@ -36,11 +36,14 @@ dataset = pybop.Dataset(
 parameter_values.update(
     {
         "Negative electrode active material volume fraction": pybop.Parameter(
-            prior=pybop.Gaussian(0.6, 0.05),
-            bounds=[0.5, 0.8],
+            distribution=pybop.Gaussian(
+                0.6,
+                0.05,
+                truncated_at=[0.5, 0.8],
+            )
         ),
         "Positive electrode active material volume fraction": pybop.Parameter(
-            prior=pybop.Gaussian(0.48, 0.05),
+            distribution=pybop.Gaussian(0.48, 0.05),
         ),
     }
 )