+++ examples/scripts/getting_started/pints_ask_tell_interface.py
@@ -29,10 +29,10 @@ dataset = pybop.Dataset(
 parameter_values.update(
     {
         "Negative electrode active material volume fraction": pybop.Parameter(
-            prior=pybop.Gaussian(0.55, 0.05),
+            distribution=pybop.Gaussian(0.55, 0.05),
         ),
         "Positive electrode active material volume fraction": pybop.Parameter(
-            prior=pybop.Gaussian(0.55, 0.05),
+            distribution=pybop.Gaussian(0.55, 0.05),
         ),
     }
 )