+++ tests/unit/test_classifier.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pybamm
 import pytest
+from scipy import stats
 
 import pybop
 
@@ -34,8 +35,7 @@ class TestClassifier:
         parameter_values.update(
             {
                 "R0 [Ohm]": pybop.Parameter(
-                    prior=pybop.Uniform(0.001, 0.1),
-                    bounds=[1e-4, 0.1],
+                    distribution=stats.uniform(loc=0.001, scale=0.1 - 0.001)
                 )
             }
         )